requirements.txt
accelerate
langchain
langchain-huggingface
langchain-openai
python-dotenv
torch
transformers